:white_check_mark: getCategories 컨트롤러 테스트 Mock방식으로 변경
