+++ application/src/main/java/com/depromeet/ahmatda/category/dto/CategoryResponse.java
@@ -1,6 +1,7 @@
 package com.depromeet.ahmatda.category.dto;
 
 import com.depromeet.ahmatda.domain.category.Category;
+import com.depromeet.ahmatda.domain.category.Emoji;
 import lombok.Builder;
 import lombok.Getter;
 
@@ -11,7 +12,7 @@ public class CategoryResponse {
     private Long id;
     private String name;
     private String type;
-    private String emoji;
+    private Emoji emoji;
 
     public static CategoryResponse createByEntity(Category category) {
         return CategoryResponse.builder()
