+++ application/src/test/java/com/depromeet/ahmatda/application/apidocs/document/ApiDocumentationTest.java
@@ -5,10 +5,12 @@ import com.depromeet.ahmatda.category.service.CategoryService;
 import com.depromeet.ahmatda.user.UserController;
 import com.depromeet.ahmatda.user.service.UserService;
 import com.fasterxml.jackson.databind.ObjectMapper;
+import org.junit.jupiter.api.BeforeEach;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.restdocs.AutoConfigureRestDocs;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
+import org.springframework.test.context.ActiveProfiles;
 import org.springframework.test.web.servlet.MockMvc;
 
 @WebMvcTest(controllers = {
@@ -17,6 +19,7 @@ import org.springframework.test.web.servlet.MockMvc;
         CategoryController.class
 })
 @AutoConfigureRestDocs
+@ActiveProfiles("test")
 public abstract class ApiDocumentationTest {
 
     @Autowired
